requirements.txt
# Core dependencies
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
plotly>=5.15.0

# Machine Learning & AI
scikit-learn>=1.3.0
tensorflow>=2.13.0
xgboost>=1.7.6
prophet>=1.1.4
arch>=5.3.0
statsmodels>=0.14.0
textblob>=0.17.1

# Data handling & visualization
openpyxl>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
fpdf>=1.7.2

# Time series specific
pandas-ta>=0.3.14b
pmdarima>=2.0.3

# API integrations
alpha_vantage>=2.3.1
newsapi-python>=0.2.7

# Development
black>=23.3.0
pytest>=7.3.1
ipykernel>=6.24.0